backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic-settings==2.7.1
sqlmodel==0.0.22
psycopg[binary]==3.2.5
redis==5.2.1
httpx==0.28.1
python-multipart==0.0.9
jinja2==3.1.4
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyJWT==2.10.1
google-api-python-client==2.146.0
google-auth==2.35.0
